Create Count_negative_elements_present_in_the_sorted_matrix_in_linear_time.c++
We start from the top-right corner of the matrix, which is the largest negative element in the matrix. Then, we traverse left if the current element is non-negative, or move down if the current element is negative. Whenever we find a negative element, we add the number of negative elements in the current row to the count, since all the elements in the current row to the left of the current element are also negative. Finally, we return the count which gives us the total number of negative elements in the matrix.