Create The K Weakest Rows in a Matrix.c
You are given an m x n binary matrix mat of 1's (representing soldiers) and 0's (representing civilians). The soldiers are positioned in front of the civilians. That is, all the 1's will appear to the left of all the 0's in each row.
Return the indices of the k weakest rows in the matrix ordered from weakest to strongest.