Create Product_of_other_numbers_of_an_array.java
Product of other numbers of an array 
Problem statement
You are given a function, int ProductArray(int" arr, int n); The function accepts an integer array 'arr' of length 'n'. Implement the function to modify the given array such that, value at each index is replaced with product of all integers of the array except for the integer on that index. 
Assumption: Array index starts from 0. n> 1 Each product operation is within the integer range. Note: Input and output arrays are of same length. 
 Example: Input: arr: 2375 Output: 105 70 30 42 Explanation: arr[0] = arr[1] * arr[2] * arr[3] = 3 *7* 5 = 105, similarly all other elements are replaced with product of all integer of the array except the integer on that particular index of array.