Create Duplicate_successive_characters.java
Few keys in the keyboards are not working properly. While typing, there are few alphabets that gets included with multiple occurrences. This is has caused huge problem
to the user. Now they have to come up with a feature that removes all these duplicates alphabets within a word. Only those alphabets have to be removed which have
occurred more than once exactly adjacent to each other.
Example:
Input: abciijklma
Output: abcjklma
Explanation: Consider the above string, it contains 2 I’s next to each other. These are the ones that need to be removed. There are also 2 a’s but they are not adjacent to
each other. Hence these should not be removed from the original string